# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rfc3339.date — Time API",
    "version": "1.0.0",
    "summary": "Strict RFC3339 time endpoints + validation + conversions (incl. exotic time standards)",
    "description": "A precise, developer-centric time API. Defaults to canonical RFC3339 UTC for /now, supports IANA time zones, strict RFC3339 validation, ISO8601 validation, and conversions across common + exotic representations (NTP, GPS, TAI, JD/MJD, Excel, HTTP-date, RFC5322). Plain text is first-class; JSON is optional via Accept header or ?json=1.\n",
    "license": {
      "name": "UNLICENSED",
      "identifier": "NOASSERTION"
    }
  },
  "servers": [
    {
      "url": "https://rfc3339.date",
      "description": "Production"
    },
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local dev (wrangler)"
    }
  ],
...
  "security": [],
  "paths": {
    "/now": {
      "get": {
        "operationId": "getNowUtc",
        "tags": [
          "Time"
        ],
        "summary": "Current time as RFC3339 (UTC)",
        "description": "Returns the current instant serialized as RFC3339 in UTC (canonical Z). Default response is text/plain.\n",
        "parameters": [
          {
            "name": "precision",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9,
              "default": 3
            },
            "description": "Fractional second digits to emit (0..9)."
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "rfc3339",
                "rfc3339sec",
                "rfc3339nano"
              ],
              "default": "rfc3339"
            },
            "description": "Convenience output format alias."
          },
          {
            "name": "json",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ],
              "default": 0
            },
            "description": "If 1, forces JSON output regardless of Accept header."
          }
        ],
        "responses": {
          "200": {
            "description": "RFC3339 timestamp",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "no-store"
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "utc": {
                    "value": "2026-02-25T19:17:03.482Z"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "now",
                    "tz",
                    "offset",
                    "precision"
                  ],
                  "properties": {
                    "now": {
                      "type": "string",
                      "description": "Timestamp string in selected representation (RFC3339)"
                    },
                    "tz": {
                      "type": "string",
                      "description": "Rendering time zone (UTC or requested IANA)"
                    },
                    "offset": {
                      "type": "string",
                      "description": "Numeric offset at that instant (e.g. Z, +01:00)"
                    },
                    "precision": {
                      "type": "integer"
                    },
                    "unix": {
                      "type": "integer",
                      "description": "Unix seconds"
                    },
                    "unixms": {
                      "type": "integer",
                      "description": "Unix milliseconds"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid input, ambiguity, or unsupported combination)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/now/{tz}": {
      "get": {
        "operationId": "getNowInZone",
        "tags": [
          "Time",
          "Timezones"
        ],
        "summary": "Current time rendered in an IANA time zone",
        "description": "Returns the current instant rendered in the requested IANA time zone, including numeric offset. Same instant as /now, different representation.\n",
        "parameters": [
          {
            "name": "tz",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "IANA time zone ID (URL-encoded), e.g. Europe%2FBerlin"
          },
          {
            "name": "precision",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9,
              "default": 3
            },
            "description": "Fractional second digits to emit (0..9)."
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "rfc3339",
                "rfc3339sec",
                "rfc3339nano"
              ],
              "default": "rfc3339"
            },
            "description": "Convenience output format alias."
          },
          {
            "name": "json",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ],
              "default": 0
            },
            "description": "If 1, forces JSON output regardless of Accept header."
          }
        ],
        "responses": {
          "200": {
            "description": "RFC3339 timestamp in requested zone (includes offset)",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "no-store"
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "berlin": {
                    "value": "2026-02-25T20:17:03.482+01:00"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "now",
                    "tz",
                    "offset",
                    "precision"
                  ],
                  "properties": {
                    "now": {
                      "type": "string",
                      "description": "Timestamp string in selected representation (RFC3339)"
                    },
                    "tz": {
                      "type": "string",
                      "description": "Rendering time zone (UTC or requested IANA)"
                    },
                    "offset": {
                      "type": "string",
                      "description": "Numeric offset at that instant (e.g. Z, +01:00)"
                    },
                    "precision": {
                      "type": "integer"
                    },
                    "unix": {
                      "type": "integer",
                      "description": "Unix seconds"
                    },
                    "unixms": {
                      "type": "integer",
                      "description": "Unix milliseconds"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid input, ambiguity, or unsupported combination)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found (unknown time zone or resource)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}